+++ ConfluenceOtherworld/src/main/java/org/confluence/mod/common/item/potion/ManaPotionItem.java
@@ -29,7 +29,7 @@ public class ManaPotionItem extends AbstractPotionItem {
     protected void apply(ItemStack itemStack, Level level, LivingEntity living) {
         if (level.isClientSide) return;
         if (living instanceof ServerPlayer serverPlayer) {
-            PlayerUtils.receiveMana(serverPlayer, () -> amount);
+            PlayerUtils.receiveMana(serverPlayer, () -> (float) amount);
             MobEffectInstance instance = serverPlayer.getEffect(ModEffects.MANA_SICKNESS);
             if (instance == null) {
                 serverPlayer.addEffect(new MobEffectInstance(ModEffects.MANA_SICKNESS, 100));
